requirements.txt
streamlit>=1.31.1
google-generativeai>=0.5.0
gspread>=6.0.0
pandas
